Align NegXplus1 bf16 fast path with conversion style
Agent-Logs-Url: https://github.com/xadupre/yet-another-onnxruntime-extensions/sessions/9d7eb8b3-9654-40cb-b376-cbda3ffe3ffc

+++ yaourt/ortops/fused_kernel/cuda/negxplus1.cu
@@ -36,7 +36,7 @@ template <> __device__ __inline__ __nv_bfloat16 _neg1plusx(const __nv_bfloat16 x
 #if __CUDA_ARCH__ < 800
   return __float2bfloat16(1.0f - __bfloat162float(x));
 #else
-  return (__nv_bfloat16)1 - x;
+  return __float2bfloat16(1.0f) - x;
 #endif
 }
